// types/equipment.ts
export interface Equipment {
  title: string
  description: string
  image: string
  link: string
  category: string
  features: string[]
  tag?: string
  badge?: string
  linkText?: string
}

export interface Epi {
  title: string
  description: string
  image: string
  link: string
  category: string
  features: string[]
  tag?: string
}

export interface Service {
  title: string
  description: string
  icon: string
} 